requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.0,<0.16
Pillow>=9.1.0,<9.2
uwsgi>=2.0.20,<2.1
